refactor: consolidate file operations
